+++ mod-data-import-converter-storage-server/src/test/java/org/folio/rest/impl/association/wrapper/MappingProfileWrapper.java
@@ -4,6 +4,10 @@ import org.folio.rest.jaxrs.model.MappingProfile;
 import org.folio.rest.jaxrs.model.Metadata;
 import org.folio.rest.jaxrs.model.UserInfo;
 
+
+/**
+ * Wrapper for {@link MappingProfile} entity
+ */
 public class MappingProfileWrapper implements ProfileWrapper<MappingProfile> {
 
   private MappingProfile mappingProfile;